+++ example/basic/main.js
@@ -6,13 +6,13 @@ let accountId;
 // Print result on screen
 function printResult(results) {
   document
-    .getElementById("result")
-    .appendChild(renderjson.set_icons("+", "-").set_show_to_level(2)(results));
+    .getElementById('result')
+    .appendChild(renderjson.set_icons('+', '-').set_show_to_level(2)(results));
 }
 
 // Get selected account's ID
 function getSelectedAccountId() {
-  return $("#account_id_mod").val();
+  return $('#account_id_mod').val();
 }
 
 // Function to handle user login
@@ -21,55 +21,80 @@ function login() {
     connectURL: constants.VEZGO_CONNECT_URL,
     clientId: constants.VEZGO_CLIENT_ID,
     baseURL: constants.VEZGO_API_URL || 'https://api.vezgo.com/v1',
-    authEndpoint: "/vezgo/auth",
+    authEndpoint: '/vezgo/auth',
     auth: {
-      headers: { Authorization: `Bearer ${$("#username").val()}` },
+      headers: { Authorization: `Bearer ${$('#username').val()}` },
     },
     // To enable "Demo" provider
-    demo: true
+    demo: true,
   });
-  $("#result").html("");
-  $("#response_heading").html(`Loading... `);
+  $('#result').html('');
+  $('#response_heading').html('Loading... ');
   user = vezgo.login();
 }
 
 // Document ready event
-$(document).ready(function () {
+$(document).ready(() => {
   // Connect button click handler
-  $("#connect").click(function () {
-    $("#connection_error strong").text("");
+  $('#connect').click(() => {
+    $('#connection_error strong').text('');
     login();
+    const multiWallet = window.location.search.includes('multi_wallet=true');
 
     user
       .connect({
+        // provider: 'binance',
+        providers: ['metamask', 'demo', 'coinbase'],
         theme: constants.VEZGO_CLIENT_THEME,
         providersPerLine: constants.VEZGO_CLIENT_PROVIDERS_PER_LINE,
         connectionType: constants.VEZGO_CONNECT_TYPE,
         // providerCategories: ['exchanges', 'wallets'],
+        multiWallet,
       })
-      .onEvent(function (event, data) {
+      .onEvent((event, data) => {
         console.log('event', event, data);
       })
-      .onConnection(function (account) {
-        console.log("connection success", account);
-        $("#response_heading").html(`Account connected successfully with ID: ${account}`);
-        $("#account_id_mod").val(account);
+      .onConnection((account) => {
+        let accounts;
+        let message;
+
+        if (multiWallet) {
+          accounts = account.accounts;
+          // eslint-disable-next-line no-param-reassign
+          message = account.message;
+        }
+
+        if (multiWallet) {
+          console.log('connection result', account);
+
+          const rows = accounts.map((acc, key) => `<tr><td>${key + 1}</td><td>${acc.account || '-'}</td><td>${acc.wallet || '-'}</td><td>${acc.message}</td></tr>`).join('');
+
+          $('#response_heading').html(message);
+          const head = '<thead><tr><th><strong>№</strong></th><th><strong>Account ID</strong></th><th><strong>Wallet</strong></th><th><strong>Status</strong></th></tr></thead>';
+
+          $('#result').html(`<table class="table border">${head}<tbody>${rows}</tbody></table>`);
+        } else {
+          console.log('connection success', account);
+          $('#response_heading').html(`Account connected successfully with ID: ${account}`);
+        }
+
+        $('#account_id_mod').val(accounts);
       })
-      .onError(function (error) {
-        console.log("connection error", error);
-        $("#response_heading").html(`Error connecting account: ${JSON.stringify(error)}`);
-        $("#connection_error strong").text(error.message);
+      .onError((error) => {
+        console.log('connection error', error);
+        $('#response_heading').html(`Error connecting account: ${JSON.stringify(error)}`);
+        $('#connection_error strong').text(error.message);
       });
   });
 
   // Reconnect button click handler
-  $("#reconnect").click(function () {
+  $('#reconnect').click(() => {
     const accountId = getSelectedAccountId();
     if (!accountId) {
-      alert("Must enter an Account ID first.");
+      alert('Must enter an Account ID first.');
       return;
     }
-    $("#connection_error strong").text("");
+    $('#connection_error strong').text('');
     login();
 
     user
@@ -78,41 +103,41 @@ $(document).ready(function () {
         providersPerLine: constants.VEZGO_CLIENT_PROVIDERS_PER_LINE,
         connectionType: constants.VEZGO_CONNECT_TYPE,
       })
-      .onEvent(function (event, data) {
+      .onEvent((event, data) => {
         console.log('event', event, data);
       })
-      .onConnection(function (account) {
-        console.log("reconnection success", account);
-        $("#response_heading").html(`Account reconnected successfully with ID: ${account}`);
-        $("#account_id_mod").val(account);
+      .onConnection((account) => {
+        console.log('reconnection success', account);
+        $('#response_heading').html(`Account reconnected successfully with ID: ${account}`);
+        $('#account_id_mod').val(account);
       })
-      .onError(function (error) {
-        console.log("reconnection error", error);
-        $("#response_heading").html(`Error reconnecting account: ${JSON.stringify(error)}`);
-        $("#connection_error strong").text(error.message);
+      .onError((error) => {
+        console.log('reconnection error', error);
+        $('#response_heading').html(`Error reconnecting account: ${JSON.stringify(error)}`);
+        $('#connection_error strong').text(error.message);
       });
   });
 
   // Get all accounts button click handler
-  $("#get_accounts").click(async function () {
+  $('#get_accounts').click(async () => {
     login();
 
     try {
       // Get all connected account
       const accounts = await user.accounts.getList();
-      $("#response_heading").html(`All accounts list of user:`);
+      $('#response_heading').html('All accounts list of user:');
       printResult(accounts);
     } catch (err) {
-      $("#response_heading").html("");
-      $("#result").html("Error:<br><code>" + err + "</code>");
+      $('#response_heading').html('');
+      $('#result').html(`Error:<br><code>${err}</code>`);
     }
   });
 
   // Account info button click handler
-  $("#get_account").click(async function () {
+  $('#get_account').click(async () => {
     accountId = getSelectedAccountId();
     if (!accountId) {
-      alert("Must enter an Account ID first.");
+      alert('Must enter an Account ID first.');
       return;
     }
     login();
@@ -120,32 +145,34 @@ $(document).ready(function () {
     try {
       // Get account's info by providing accountId
       const account = await user.accounts.getOne(accountId);
-      $("#response_heading").html(`Information for account with ID: ${accountId}`);
+      $('#response_heading').html(`Information for account with ID: ${accountId}`);
       printResult(account);
     } catch (err) {
-      $("#response_heading").html("");
-      $("#result").html("Error:<br><code>" + err + "</code>");
+      $('#response_heading').html('');
+      $('#result').html(`Error:<br><code>${err}</code>`);
     }
   });
 
   // Get transactions button click handler
-  $("#get_transactions").click(async function () {
+  $('#get_transactions').click(async () => {
     const accountId = getSelectedAccountId();
     if (!accountId) {
-      alert("Must connect an account first");
+      alert('Must connect an account first');
       return;
     }
     login();
 
     try {
       const transactions = [];
-      const from = "2012-01-01";
-      let last = "";
+      const from = '2012-01-01';
+      let last = '';
       const limit = 50;
 
       while (true) {
         // Getting all transactions by accountId
-        const page = await user.transactions.getList({ accountId, from, last, limit });
+        const page = await user.transactions.getList({
+          accountId, from, last, limit,
+        });
 
         if (!page.length) break;
 
@@ -155,22 +182,21 @@ $(document).ready(function () {
         last = page[page.length - 1].id;
       }
 
-      $("#response_heading").html(`Transactions for account with ID: ${accountId}`);
-      if(transactions.length == 0){
-        $("#result").html("No Transactions found!");
-
+      $('#response_heading').html(`Transactions for account with ID: ${accountId}`);
+      if (transactions.length == 0) {
+        $('#result').html('No Transactions found!');
       }
     } catch (err) {
-      $("#response_heading").html("");
-      $("#result").html("Error:<br><code>" + err + "</code>");
+      $('#response_heading').html('');
+      $('#result').html(`Error:<br><code>${err}</code>`);
     }
   });
 
   // Sync account button click handler
-  $("#sync_account").click(async function () {
+  $('#sync_account').click(async () => {
     accountId = getSelectedAccountId();
     if (!accountId) {
-      alert("Must enter an Account ID first.");
+      alert('Must enter an Account ID first.');
       return;
     }
     login();
@@ -178,18 +204,18 @@ $(document).ready(function () {
     try {
       // Sync account by  accountId
       await user.accounts.sync(accountId);
-      $("#response_heading").html(`Sync requested for ID: ${accountId}`);
+      $('#response_heading').html(`Sync requested for ID: ${accountId}`);
     } catch (err) {
-      $("#response_heading").html("");
-      $("#result").html("Error:<br><code>" + err + "</code>");
+      $('#response_heading').html('');
+      $('#result').html(`Error:<br><code>${err}</code>`);
     }
   });
 
   // Remove account button click handler
-  $("#remove_account").click(async function () {
+  $('#remove_account').click(async () => {
     accountId = getSelectedAccountId();
     if (!accountId) {
-      alert("Must enter an Account ID first.");
+      alert('Must enter an Account ID first.');
       return;
     }
     login();
@@ -197,10 +223,10 @@ $(document).ready(function () {
     try {
       // Remove account by  accountId
       await user.accounts.remove(accountId);
-      $("#response_heading").html(`Removed account for ID: ${accountId}`);
+      $('#response_heading').html(`Removed account for ID: ${accountId}`);
     } catch (err) {
-      $("#response_heading").html("");
-      $("#result").html("Error:<br><code>" + err + "</code>");
+      $('#response_heading').html('');
+      $('#result').html(`Error:<br><code>${err}</code>`);
     }
   });
 });