+++ src/api.js
@@ -183,6 +183,7 @@ class API {
       theme,
       providersPerLine,
       features,
+      multiWallet,
     } = options;
     const { clientId, connectURL } = this.config;
 
@@ -229,6 +230,7 @@ class API {
       theme: ['light', 'dark'].includes(theme) ? theme : 'light',
       providers_per_line: (providersPerLine && ['1', '2'].includes(providersPerLine.toString())) ? providersPerLine.toString() : '2',
       features,
+      multi_wallet: multiWallet,
     };
 
     // Cleanup blank params
@@ -300,6 +302,7 @@ class API {
           providersPerLine,
           syncNfts,
           features,
+          multiWallet,
         } = options;
         const { url, token } = await this.getConnectData({
           provider,
@@ -311,6 +314,7 @@ class API {
           providersPerLine,
           syncNfts,
           features,
+          multiWallet,
         });
 
         this.iframe = appendVezgoIframe();
@@ -357,6 +361,7 @@ class API {
       case 'success': {
         // Connection success
         this._triggerCallback(CALLBACK_CONNECTION, result.account);
+
         break;
       }
 
@@ -423,7 +428,7 @@ class API {
     if (this.form) this.form.remove();
   }
 
-  _triggerCallback(callback, payload = {}) {
+  _triggerCallback(callback, payload) {
     // onConnection and onError callbacks are only triggered once per widget session.
     if ([CALLBACK_CONNECTION, CALLBACK_ERROR].includes(callback)) {
       // Mark widget as inactive so it's cleaned up by the doneWatcher.
@@ -439,7 +444,10 @@ class API {
     }
 
     if (callback === CALLBACK_EVENT && this[callback]) {
-      this[callback](payload.event, payload.data);
+      const event = payload ? payload.event : null;
+      const data = payload ? payload.data : null;
+
+      this[callback](event, data);
     }
   }